[FIX] stock_account: Search valuation method on product
Since the refactoring of making the product category optional. There
still a lot of domain in the code that check if [('product.categ_id..valuation', '=', ...)]
Howerver it's not correct anymore since there is no fall back if
the category is not set.

In order to make it easier. A new search function has been added on
product to directly search on the category and use a fallback on the
company default if the categ is not set or the valuation is not set
on the company

X-original-commit: a799a3ee233da93973a8ba2eeb5944de8ce53945
Part-of: odoo/odoo#230000
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>